Remove old types, rename new types with clearer names
