+++ guice/hibernate/src/main/java/com/peterphi/std/guice/hibernate/webquery/impl/HQLFragment.java
@@ -3,12 +3,12 @@ package com.peterphi.std.guice.hibernate.webquery.impl;
 import java.util.List;
 import java.util.stream.Collectors;
 
-public class HSQLFragment
+public class HQLFragment
 {
 	private final String expr;
 
 
-	public HSQLFragment(final String expr)
+	public HQLFragment(final String expr)
 	{
 		this.expr = expr;
 	}
@@ -19,10 +19,10 @@ public class HSQLFragment
 	}
 
 
-	public static HSQLFragment combine(final List<HSQLFragment> fragments,
-	                                   final String prefix,
-	                                   final String separator,
-	                                   final String suffix)
+	public static HQLFragment combine(final List<HQLFragment> fragments,
+	                                  final String prefix,
+	                                  final String separator,
+	                                  final String suffix)
 	{
 		// Combine all the expressions
 		final String expr = prefix +
@@ -33,6 +33,6 @@ public class HSQLFragment
 				                    .filter(f -> f != null)
 				                    .collect(Collectors.joining(separator)) + suffix;
 
-		return new HSQLFragment(expr);
+		return new HQLFragment(expr);
 	}
 }
