+++ guice/hibernate/src/main/java/com/peterphi/std/guice/hibernate/webquery/impl/QEntity.java
@@ -1,7 +1,5 @@
 package com.peterphi.std.guice.hibernate.webquery.impl;
 
-import com.google.common.cache.Cache;
-import com.google.common.cache.CacheBuilder;
 import com.peterphi.std.guice.database.annotation.SearchFieldAlias;
 import com.peterphi.std.guice.restclient.jaxb.webqueryschema.WQEntitySchema;
 import org.apache.commons.lang.StringUtils;
@@ -19,7 +17,6 @@ import java.util.Arrays;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.HashMap;
-import java.util.HashSet;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
@@ -36,13 +33,6 @@ public class QEntity
 	private Map<String, QRelation> relations = new HashMap<>();
 
 	private List<QEntity> descendants = Collections.emptyList();
-
-	private final Cache<String, QPropertyPathBuilder> builderCache = CacheBuilder
-			                                                                 .newBuilder()
-			                                                                 .weakKeys()
-			                                                                 .weakValues()
-			                                                                 .initialCapacity(0)
-			                                                                 .build();
 
 
 	public QEntity(Class<?> clazz)
@@ -290,86 +280,6 @@ public class QEntity
 			                                   this.clazz.getSimpleName() +
 			                                   ", expected one of " +
 			                                   properties.keySet());
-	}
-
-
-	public QPropertyPathBuilder getPath(final String path)
-	{
-		QPropertyPathBuilder cached = builderCache.getIfPresent(path);
-
-		if (cached == null)
-		{
-			final QPropertyPathBuilder builder = new QPropertyPathBuilder();
-
-			cached = getPath(builder, path);
-
-			builderCache.put(path, cached);
-		}
-
-		return cached;
-	}
-
-
-	public QPropertyPathBuilder getPath(final QPropertyPathBuilder builder, String path)
-	{
-		final int firstDot = path.indexOf('.');
-
-		final boolean terminal = (firstDot == -1);
-
-		final String head = terminal ? path : path.substring(0, firstDot);
-		final String tail = terminal ? null : path.substring(firstDot + 1);
-
-		if (aliases.containsKey(head))
-		{
-			final String newPath;
-
-			if (terminal)
-				newPath = getAlias(head);
-			else
-				newPath = getAlias(head) + "." + tail;
-
-			return getPath(builder, newPath);
-		}
-		else if (hasProperty(head))
-		{
-			if (tail != null)
-				throw new IllegalArgumentException("Found property " + head + " but there are other path components: " + tail);
-
-			builder.append(getProperty(head));
-
-			return builder;
-		}
-		else if (hasRelation(head))
-		{
-			final QRelation relation = getRelation(head);
-
-			builder.append(getRelation(head));
-
-			return relation.getEntity().getPath(builder, tail);
-		}
-		else
-		{
-			final Set<String> expected = new HashSet<>(getPropertyNames());
-			expected.addAll(getRelationNames());
-			expected.addAll(getAliasNames());
-
-			throw new IllegalArgumentException("Relationship path error: got " + head + ", expected one of: " + expected);
-		}
-	}
-
-
-	/**
-	 * Retrieve the resolution of a defined alias on this entity. If there is no such destination alias then return null<br />
-	 * This is designed to allow underlying database schema changes without changing the query API exposed to users
-	 *
-	 * @param name
-	 * 		some aliased name (e.g. "assetId")
-	 *
-	 * @return some alias destination (e.g. "asset.id")
-	 */
-	public String getAlias(String name)
-	{
-		return aliases.get(name);
 	}
