+++ guice/hibernate/src/main/java/com/peterphi/std/guice/hibernate/webquery/impl/QFunction.java
@@ -4,5 +4,5 @@ import com.peterphi.std.guice.hibernate.webquery.HQLEncodingContext;
 
 public interface QFunction
 {
-	HSQLFragment encode(HQLEncodingContext ctx);
+	HQLFragment encode(HQLEncodingContext ctx);
 }
