+++ guice/hibernate/src/main/java/com/peterphi/std/guice/hibernate/webquery/impl/functions/BinaryFunction.java
@@ -1,7 +1,7 @@
 package com.peterphi.std.guice.hibernate.webquery.impl.functions;
 
 import com.peterphi.std.guice.hibernate.webquery.HQLEncodingContext;
-import com.peterphi.std.guice.hibernate.webquery.impl.HSQLFragment;
+import com.peterphi.std.guice.hibernate.webquery.impl.HQLFragment;
 import com.peterphi.std.guice.hibernate.webquery.impl.QFunction;
 import com.peterphi.std.guice.hibernate.webquery.impl.QPropertyRef;
 
@@ -21,8 +21,8 @@ class BinaryFunction implements QFunction
 
 
 	@Override
-	public HSQLFragment encode(final HQLEncodingContext ctx)
+	public HQLFragment encode(final HQLEncodingContext ctx)
 	{
-		return new HSQLFragment(property.toHqlPath() + " " + operator + " " + ctx.createPropertyPlaceholder(value));
+		return new HQLFragment(property.toHqlPath() + " " + operator + " " + ctx.createPropertyPlaceholder(value));
 	}
 }
